+++ src/main/java/nextstep/courses/infrastructure/course/JdbcCourseRepository.java
@@ -1,4 +1,4 @@
-package nextstep.courses.infrastructure;
+package nextstep.courses.infrastructure.course;
 
 import nextstep.courses.domain.course.Course;
 import org.springframework.jdbc.core.JdbcOperations;
